msvc: Add support for libav/ffmpeg compiled with msvc.
By default when building ffms2 with msvc it will assume that
libav/ffmpeg were also built with msvc.  This avoids the various
problems of linking against mingw static libraries in msvc and makes the
world a little saner.

Also includes support for opencore-amr built with msvc.  Support for
libraries other than opencore-amr may be added later.

Removes support for the autodetection of libraries which can now be
error prone with the additional msvc support in libav/ffmpeg.
Additional libraries must now be manually enabled in msvc-config.h.

Originally committed to SVN as r714.